Soft/TCP_client_and_Server/code/requirements.txt
Package                Version
---------------------- ---------
cflib                  0.1.17.1
dynamixel-sdk          3.7.31
libusb-package         1.0.24.2
numpy                  1.22.1
opencv-python-headless 4.5.5.62
packaging              21.3
pandas                 1.4.0
pip                    20.3.4
pkg-resources          0.0.0
pyparsing              3.0.7
PyQt5-sip              12.9.0
pyserial               3.5
python-dateutil        2.8.2
python-dynamixel       -UNKNOWN-
pytz                   2021.3
pyusb                  1.2.1
QtPy                   2.0.0
scipy                  1.7.3
setuptools             44.1.1
six                    1.16.0

